use resource name for IP address on DNS entry
The IP address for the OpsManager DNS entry was being determine via the NAT IP address assigned to the VM. This worked™, but caused a strange dependency graph.

The DNS entry depends on the IP address resource now.

This is the first pass at creating the infrastructure only terraform. With the next step being the OpsMan VM being optional, but have all the required ref-arch supported.